devtools/server/tests/mochitest/test_inspector-reload.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gInspectee = null;
let gWalker = null;

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  const inspector = await target.getInspector();
  gInspectee = doc;
  const walker = inspector.walker;
  gWalker = await inspector.getWalker();
  ok(walker === gWalker, "getWalker() twice should return the same walker.");
  runNextTest();
});

addTest(function testReload() {
  const oldRootID = gWalker.rootNode.actorID;
  // Load a node to populate the tree a bit.
  promiseDone(gWalker.querySelector(gWalker.rootNode, "#a").then(front => {
    gInspectee.defaultView.location.reload();
    return waitForMutation(gWalker, isNewRoot);
  }).then(() => {
    ok(gWalker.rootNode.actorID != oldRootID, "Root node should have changed.");
  }).then(() => {
    // Make sure we can still access the document
    return gWalker.querySelector(gWalker.rootNode, "#a");
  }).then(front => {
    ok(front.actorID, "Got a new actor ID");
  }).then(runNextTest));
});

addTest(function cleanup() {
  gWalker = null;
  gInspectee = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
